// src/styles/makingColorsTheme/interfaces.tsx
export interface IColors {
  default: ImakeColorsTone
  main: ImakeColorsTone
  second: ImakeColorsTone
  grey: ImakeColorsTone
  greyAzure?: ImakeColorsTone
  limeGreen?: ImakeColorsTone
  green?: ImakeColorsTone
  leaf?: ImakeColorsTone
  paleGreen?: ImakeColorsTone
  teal?: ImakeColorsTone
  cyan?: ImakeColorsTone
  blue?: ImakeColorsTone
  greyBlue?: ImakeColorsTone
  greyPurple?: ImakeColorsTone
  purple?: ImakeColorsTone
  wine?: ImakeColorsTone
  paleRed?: ImakeColorsTone
  pink?: ImakeColorsTone
  red?: ImakeColorsTone
  strongRed?: ImakeColorsTone
  brown?: ImakeColorsTone
  orange?: ImakeColorsTone
  strongYellow?: ImakeColorsTone
  yellow?: ImakeColorsTone
  gold?: ImakeColorsTone
  cream?: ImakeColorsTone
}

type typeOfColor = string | undefined
export interface ImakeColorsTone {
  c025?: typeOfColor,
  c050?: typeOfColor,
  c075?: typeOfColor,
  c100?: typeOfColor,
  c125?: typeOfColor,
  c150?: typeOfColor,
  c175?: typeOfColor,
  c200?: typeOfColor,
  c225?: typeOfColor,
  c250?: typeOfColor,
  c275?: typeOfColor,
  c300?: typeOfColor,
  c325?: typeOfColor,
  c350?: typeOfColor,
  c400?: typeOfColor,
  c450?: typeOfColor,
  c500?: typeOfColor,
  c550?: typeOfColor,
  c600?: typeOfColor,
  c650?: typeOfColor,
  c675?: typeOfColor,
  c700?: typeOfColor,
  c725?: typeOfColor,
  c750?: typeOfColor,
  c775?: typeOfColor,
  c800?: typeOfColor,
  c825?: typeOfColor,
  c850?: typeOfColor,
  c875?: typeOfColor,
  c900?: typeOfColor,
  c925?: typeOfColor,
  c950?: typeOfColor,
  c975?: typeOfColor,
}